pyproject.toml
[project]
name = "integra"
version = "0.1.0"
description = "Master/orchestrator agent with secure data lake, Telegram HIL, and extensible tool registry"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.52.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "python-telegram-bot>=22.0",
    "pyrage>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.13"
strict = true

[[tool.mypy.overrides]]
module = ["pyrage", "pyrage.*"]
ignore_missing_imports = true
follow_imports = "skip"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
